HashMap : Added try-catch and screenshots
